requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
djangorestframework==3.16.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
